load the training results in to the cloud and create a function to load the model
