+++ .env
@@ -2,3 +2,22 @@
 
 SPOON_API_KEY = 5cf4d9752bea4c38b962e643ca227e27
 ROBO_API_KEY = qhSk7QdaM3p1YIzIdrPZ
+
+DATA_SIZE=1k
+CHUNK_SIZE=200
+MODEL_TARGET=local
+
+# GCP Project
+GCP_PROJECT_WAGON=wagon-public-datasets
+GCP_PROJECT= "precise-bank-374911"
+GCP_REGION=europe-west1
+
+# Cloud Storage
+BUCKET_NAME=icook
+
+# BigQuery
+BQ_REGION=EU
+BQ_DATASET="question"
+
+# Compute Engine
+INSTANCE= "question"
