+++ icook/params.py
@@ -1,0 +1,11 @@
+import os
+
+SPOON_API_KEY = os.environ.get('SPOON_API_KEY')
+ROBO_API_KEY = os.environ.get('ROBO_API_KEY')
+GCP_PROJECT = os.environ.get("GCP_PROJECT")
+GCP_PROJECT_WAGON = os.environ.get("GCP_PROJECT_WAGON")
+GCP_REGION = os.environ.get("GCP_REGION")
+BUCKET_NAME = os.environ.get("BUCKET_NAME")
+GCR_IMAGE = os.environ.get("GCR_IMAGE")
+GCR_REGION = os.environ.get("GCR_REGION")
+GCR_MEMORY = os.environ.get("GCR_MEMORY")
